HAP-1093 - Declarative syntax for Jenkins Artifactory Plugin
