+++ src/main/java/org/jfrog/hudson/pipeline/declarative/steps/maven/MavenDeployerStep.java
@@ -21,7 +21,37 @@ public class MavenDeployerStep extends MavenDeployerResolver {
     }
 
     @DataBoundSetter
-    public void setDeployEvenIfUnstable(String deployEvenIfUnstable) {
+    public void setIncludePattern(String includePattern) {
+        mavenDeployer.getArtifactDeploymentPatterns().addInclude(includePattern);
+    }
+
+    @DataBoundSetter
+    public void setExcludePattern(String excludePattern) {
+        mavenDeployer.getArtifactDeploymentPatterns().addExclude(excludePattern);
+    }
+
+    @DataBoundSetter
+    public void setIncludeEnvVars(boolean includeEnvVars) {
+        mavenDeployer.setIncludeEnvVars(includeEnvVars);
+    }
+
+    @DataBoundSetter
+    public void setCustomBuildName(String customBuildName) {
+        mavenDeployer.setCustomBuildName(customBuildName);
+    }
+
+    @DataBoundSetter
+    public void setProperties(String properties) {
+        buildDataFile.put("properties", properties);
+    }
+
+    @DataBoundSetter
+    public void setDeployArtifacts(boolean deployArtifacts) {
+        mavenDeployer.setDeployArtifacts(deployArtifacts);
+    }
+
+    @DataBoundSetter
+    public void setDeployEvenIfUnstable(boolean deployEvenIfUnstable) {
         mavenDeployer.setDeployEvenIfUnstable(deployEvenIfUnstable);
     }
