+++ src/main/java/org/jfrog/hudson/pipeline/scripted/steps/ArtifactoryMavenBuild.java
@@ -1,4 +1,4 @@
-package org.jfrog.hudson.pipeline.steps;
+package org.jfrog.hudson.pipeline.scripted.steps;
 
 import com.google.inject.Inject;
 import hudson.EnvVars;
@@ -11,55 +11,43 @@ import org.jenkinsci.plugins.workflow.steps.AbstractStepDescriptorImpl;
 import org.jenkinsci.plugins.workflow.steps.AbstractStepImpl;
 import org.jenkinsci.plugins.workflow.steps.AbstractSynchronousNonBlockingStepExecution;
 import org.jenkinsci.plugins.workflow.steps.StepContextParameter;
-import org.jfrog.hudson.pipeline.executors.GradleExecutor;
-import org.jfrog.hudson.pipeline.types.GradleBuild;
+import org.jfrog.hudson.pipeline.executors.MavenExecutor;
+import org.jfrog.hudson.pipeline.types.MavenBuild;
 import org.jfrog.hudson.pipeline.types.buildInfo.BuildInfo;
 import org.kohsuke.stapler.DataBoundConstructor;
 
 /**
  * Created by Tamirh on 04/08/2016.
  */
-public class ArtifactoryGradleBuild extends AbstractStepImpl {
+public class ArtifactoryMavenBuild extends AbstractStepImpl {
 
-    private GradleBuild gradleBuild;
-    private String tasks;
-    private String buildFile;
-    private String rootDir;
-    private String switches;
+    private MavenBuild mavenBuild;
+    private String goal;
+    private String pom;
     private BuildInfo buildInfo;
 
     @DataBoundConstructor
-    public ArtifactoryGradleBuild(GradleBuild gradleBuild, String rootDir, String buildFile, String tasks, String switches, BuildInfo buildInfo) {
-        this.gradleBuild = gradleBuild;
-        this.tasks = tasks;
-        this.rootDir = rootDir;
-        this.buildFile = buildFile;
-        this.switches = switches;
+    public ArtifactoryMavenBuild(MavenBuild mavenBuild, String pom, String goals, BuildInfo buildInfo) {
+        this.mavenBuild = mavenBuild;
+        this.goal = goals == null ? "" : goals;
+        this.pom = pom == null ? "" : pom;
         this.buildInfo = buildInfo;
     }
 
-    private GradleBuild getGradleBuild() {
-        return gradleBuild;
+    private MavenBuild getMavenBuild() {
+        return mavenBuild;
     }
 
-    private String getSwitches() {
-        return switches;
+    private String getGoal() {
+        return goal;
     }
 
-    private String getTasks() {
-        return tasks;
-    }
-
-    private String getBuildFile() {
-        return buildFile;
+    private String getPom() {
+        return pom;
     }
 
     private BuildInfo getBuildInfo() {
         return buildInfo;
-    }
-
-    private String getRootDir() {
-        return rootDir;
     }
 
     public static class Execution extends AbstractSynchronousNonBlockingStepExecution<BuildInfo> {
@@ -75,7 +63,7 @@ public class ArtifactoryGradleBuild extends AbstractStepImpl {
         private transient Launcher launcher;
 
         @Inject(optional = true)
-        private transient ArtifactoryGradleBuild step;
+        private transient ArtifactoryMavenBuild step;
 
         @StepContextParameter
         private transient FilePath ws;
@@ -85,9 +73,10 @@ public class ArtifactoryGradleBuild extends AbstractStepImpl {
 
         @Override
         protected BuildInfo run() throws Exception {
-            GradleExecutor gradleExecutor = new GradleExecutor(build, step.getGradleBuild(), step.getTasks(), step.getBuildFile(), step.getRootDir(), step.getSwitches(), step.getBuildInfo(), env, ws, listener, launcher);
-            gradleExecutor.execute();
-            return gradleExecutor.getBuildInfo();
+            MavenBuild mavenBuild = step.getMavenBuild();
+            MavenExecutor mavenExecutor = new MavenExecutor(listener, launcher, build, ws, env, mavenBuild, step.getPom(), step.getGoal(), step.getBuildInfo());
+            mavenExecutor.execute();
+            return mavenExecutor.getBuildInfo();
         }
     }
 
@@ -95,17 +84,17 @@ public class ArtifactoryGradleBuild extends AbstractStepImpl {
     public static final class DescriptorImpl extends AbstractStepDescriptorImpl {
 
         public DescriptorImpl() {
-            super(ArtifactoryGradleBuild.Execution.class);
+            super(ArtifactoryMavenBuild.Execution.class);
         }
 
         @Override
         public String getFunctionName() {
-            return "ArtifactoryGradleBuild";
+            return "artifactoryMavenBuild";
         }
 
         @Override
         public String getDisplayName() {
-            return "run Artifactory gradle";
+            return "run Artifactory maven";
         }
 
         @Override
